requirements.txt
Flask
python-dotenv
tqdm
